feat: started work on lambda terraform file
