+++ .gitignore
@@ -1,1 +1,5 @@
 venv
+.env
+python/*
+layer_content.zip
+.terraform
