+++ modules/lambda/variables.tf
@@ -1,0 +1,3 @@
+variable "webapp_public_ip" {
+  type = string
+}
