+++ requirements.txt
@@ -1,0 +1,12 @@
+boto3==1.34.144
+botocore==1.34.144
+certifi==2024.7.4
+clickhouse-connect==0.7.16
+jmespath==1.0.1
+lz4==4.3.3
+python-dateutil==2.9.0.post0
+pytz==2024.1
+s3transfer==0.10.2
+six==1.16.0
+urllib3==2.2.2
+zstandard==0.23.0
